requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.2.0
numpy>=1.24.0
torch>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tensorboard>=2.14.0
pandas>=2.0.0
